backend/requirements.txt
# ============================================================
#  OMNIGENIUS — Python dependencies
#  Pin exact versions for reproducible Lambda builds
#  Update with: pip-compile requirements.in
# ============================================================

# ── Web framework ──────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6     # Local dev server
mangum==0.19.0                # FastAPI → AWS Lambda adapter

# ── Config ─────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ── Auth (Clerk JWT verification) ──────────────────────────
python-jose[cryptography]==3.3.0
httpx==0.27.2                 # Async HTTP — used for Clerk API calls

# ── Supabase ───────────────────────────────────────────────
supabase==2.9.0
asyncpg==0.29.0               # Async Postgres driver (direct DB access)
sqlalchemy[asyncio]==2.0.35   # ORM for complex queries

# ── Stripe ─────────────────────────────────────────────────
stripe==11.1.0

# ── Redis (rate limiting + caching) ────────────────────────
redis[hiredis]==5.1.1

# ── AWS ────────────────────────────────────────────────────
boto3==1.35.36                # S3, SSM, Lambda invoke

# ── File handling ──────────────────────────────────────────
python-multipart==0.0.12     # FastAPI file uploads
Pillow==10.4.0               # Image resize for avatars
pdfplumber==0.11.4           # PDF text extraction

# ── Streaming (SSE) ────────────────────────────────────────
sse-starlette==2.1.3

# ── Observability ──────────────────────────────────────────
sentry-sdk[fastapi]==2.16.0
structlog==24.4.0

# ── Testing ────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2                # Also used in tests as async test client

# ── Flask (auth, billing, admin, webhooks, all non-inference routes) ──
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-redis==0.4.0
flask-limiter==3.8.0
apig-wsgi==2.18.0             # Flask WSGI → AWS Lambda adapter
